# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Feedora API - AI-Powered Real-Time Surplus Food Donation Network
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type UserRegistrationRole = typeof UserRegistrationRole[keyof typeof UserRegistrationRole];


export const UserRegistrationRole = {
  donor: 'donor',
  ngo: 'ngo',
  volunteer: 'volunteer',
} as const;

export interface UserRegistration {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role: UserRegistrationRole;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  organizationName?: string | null;
  /** @nullable */
  address?: string | null;
}

export interface UserLogin {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  donor: 'donor',
  ngo: 'ngo',
  volunteer: 'volunteer',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  organizationName?: string | null;
  /** @nullable */
  address?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type FoodPostInputFoodType = typeof FoodPostInputFoodType[keyof typeof FoodPostInputFoodType];


export const FoodPostInputFoodType = {
  cooked_meal: 'cooked_meal',
  raw_produce: 'raw_produce',
  packaged_food: 'packaged_food',
  bakery: 'bakery',
  dairy: 'dairy',
  beverages: 'beverages',
  other: 'other',
} as const;

export type FoodPostInputUnit = typeof FoodPostInputUnit[keyof typeof FoodPostInputUnit];


export const FoodPostInputUnit = {
  kg: 'kg',
  portions: 'portions',
  boxes: 'boxes',
  liters: 'liters',
  items: 'items',
} as const;

export interface FoodPostInput {
  foodType: FoodPostInputFoodType;
  quantity: number;
  unit: FoodPostInputUnit;
  /** Minutes until food expires */
  expiryMinutes: number;
  address: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  servings?: number | null;
}

export type FoodPostStatus = typeof FoodPostStatus[keyof typeof FoodPostStatus];


export const FoodPostStatus = {
  pending: 'pending',
  claimed: 'claimed',
  in_transit: 'in_transit',
  delivered: 'delivered',
  cancelled: 'cancelled',
  expired: 'expired',
} as const;

export interface FoodPost {
  id: number;
  donorId: number;
  donorName: string;
  foodType: string;
  quantity: number;
  unit: string;
  expiryMinutes: number;
  address: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  servings?: number | null;
  status: FoodPostStatus;
  /** @nullable */
  claimedByNgoName?: string | null;
  createdAt: string;
  expiresAt?: string;
}

export interface FoodPostList {
  posts: FoodPost[];
  total: number;
}

export type ClaimStatus = typeof ClaimStatus[keyof typeof ClaimStatus];


export const ClaimStatus = {
  claimed: 'claimed',
  picked_up: 'picked_up',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface Claim {
  id: number;
  postId: number;
  ngoId: number;
  /** @nullable */
  ngoName?: string | null;
  /** @nullable */
  volunteerId?: number | null;
  /** @nullable */
  volunteerName?: string | null;
  status: ClaimStatus;
  claimedAt: string;
  /** @nullable */
  pickedUpAt?: string | null;
  /** @nullable */
  deliveredAt?: string | null;
  foodPost?: FoodPost;
}

export interface ClaimList {
  claims: Claim[];
}

export interface Ngo {
  id: number;
  userId: number;
  organizationName: string;
  address: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  capacity?: number | null;
  reliabilityScore: number;
  totalClaimsCompleted: number;
  createdAt: string;
}

export interface NgoList {
  ngos: Ngo[];
}

export interface NgoMatch {
  ngo: Ngo;
  /** AI matching score 0-100 */
  score: number;
  /** Human-readable reason for the match */
  reason: string;
}

export interface NgoMatchList {
  matches: NgoMatch[];
}

export interface ImpactSummary {
  totalMealsSaved: number;
  totalDonations: number;
  totalDelivered: number;
  totalNgos: number;
  totalDonors: number;
  totalVolunteers: number;
  co2AvoidedKg: number;
  activePosts: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  post_created: 'post_created',
  post_claimed: 'post_claimed',
  pickup_done: 'pickup_done',
  delivery_done: 'delivery_done',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  description: string;
  /** @nullable */
  donorName?: string | null;
  /** @nullable */
  ngoName?: string | null;
  /** @nullable */
  foodType?: string | null;
  /** @nullable */
  servings?: number | null;
  createdAt: string;
}

export interface ActivityList {
  activities: ActivityItem[];
}

export interface FoodTypeBreakdown {
  foodType: string;
  count: number;
  servings: number;
}

export interface FoodTypeBreakdownList {
  breakdown: FoodTypeBreakdown[];
}

export interface Volunteer {
  id: number;
  userId: number;
  name: string;
  /** @nullable */
  phone?: string | null;
  totalDeliveries: number;
  createdAt: string;
}

export interface VolunteerList {
  volunteers: Volunteer[];
}

export type ListFoodPostsParams = {
status?: ListFoodPostsStatus;
limit?: number;
offset?: number;
};

export type ListFoodPostsStatus = typeof ListFoodPostsStatus[keyof typeof ListFoodPostsStatus];


export const ListFoodPostsStatus = {
  pending: 'pending',
  claimed: 'claimed',
  in_transit: 'in_transit',
  delivered: 'delivered',
  cancelled: 'cancelled',
  expired: 'expired',
} as const;

export type GetRecentActivityParams = {
limit?: number;
};