# lib/api-zod/src/generated/types/foodPostInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Feedora API - AI-Powered Real-Time Surplus Food Donation Network
 * OpenAPI spec version: 0.1.0
 */
import type { FoodPostInputFoodType } from './foodPostInputFoodType';
import type { FoodPostInputUnit } from './foodPostInputUnit';

export interface FoodPostInput {
  foodType: FoodPostInputFoodType;
  quantity: number;
  unit: FoodPostInputUnit;
  /** Minutes until food expires */
  expiryMinutes: number;
  address: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  servings?: number | null;
}