# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Feedora API - AI-Powered Real-Time Surplus Food Donation Network
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  organizationName?: string | null;
  /** @nullable */
  address?: string | null;
  createdAt: Date;
}