# lib/api-zod/src/generated/types/userRegistration.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Feedora API - AI-Powered Real-Time Surplus Food Donation Network
 * OpenAPI spec version: 0.1.0
 */
import type { UserRegistrationRole } from './userRegistrationRole';

export interface UserRegistration {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role: UserRegistrationRole;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  organizationName?: string | null;
  /** @nullable */
  address?: string | null;
}